requirements.txt
streamlit
opencv-python
dlib
numpy
Pillow
